# api/oas-doc.yaml
openapi: 3.0.1
info:
  title: API REST Node
  description: API REST Node
  termsOfService: ''
  contact:
    email: jgonzalez@teceng-gaming.com.ar
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: ''
servers:
  - url: 'http://localhost:8080'
  - url: ''
tags:
  - name: counter
    description: Counters managment
    externalDocs:
      description: ''
      url: ''
  - name: event
    description: Event managment
  - name: user
    description: Operations about user
    externalDocs:
      description: ''
      url: ''
paths:
  /counter:
    post:
      tags:
        - counter
      summary: Envío de contadores
      operationId: addCounter
      requestBody:
        description: Agregado de un nuevo contador
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Counter'
          application/xml:
            schema:
              $ref: '#/components/schemas/Counter'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Counter'
              example:
                egmiid : 17
                ci: 350
                co: 10
                jp: 200
                jj: 10
                drop: 0
                cc: 0
                fecha: "2019-11-04T18:15:32.673Z"  
                sala_id: 7
        '405':
          description: Invalid input
          content: {}
        '401':
          description: Invalid token
          content: {}
      security:
        - bearerAuth: []
      x-codegen-request-body-name: body
    get:
      tags:
        - counter
      summary: Obtención de contadores
      description: ''
      operationId: getCounter
      parameters:
        - name: fecha_desde
          in: query
          description: Desde esta fecha
          required: true
          style: form
          explode: true
          schema:
            type: string
            items:
              type: string
        - name: fecha_hasta
          in: query
          description: Hasta esta fecha
          required: true
          style: form
          explode: true
          schema:
            type: string
            items:
              type: string
        - name: egmid
          in: query
          description: ID de EGM
          required: false
          style: form
          explode: true
          schema:
            type: integer
            items:
              type: integer
      responses:
        '200':
          description: successful operation
          content:
            # application/xml:
            #   schema:
            #     type: array
            #     items:
            #       $ref: '#/components/schemas/Counter'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Counter'
              example:
                  _id: "5dbc7cec1fa73877d5e4c7c7"
                  egmiid : 17
                  ci: 350
                  co: 10
                  jp: 200
                  jj: 10
                  drop: 0
                  cc: 0
                  fecha: "2019-11-04T18:15:32.673Z"  
                  sala_id: 7
                  __v: 0
        '400':
          description: Invalid status value
          content: {}
        '401':
          description: Invalid token
          content: {}
      security:
        - bearerAuth: []
  /event:
    get:
      tags:
        - event
      summary: Consulta de eventos
      description: ''
      operationId: getEvent
      parameters:
        - name: fecha_desde
          in: query
          description: Desde esta fecha
          required: true
          style: form
          explode: true
          schema:
            type: string
            items:
              type: string
        - name: fecha_hasta
          in: query
          description: Hasta esta fecha
          required: true
          style: form
          explode: true
          schema:
            type: string
            items:
              type: string
        - name: egmid
          in: query
          description: ID de EGM
          required: false
          style: form
          explode: true
          schema:
            type: integer
            items:
              type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
              example:
                  _id: "5dbc7cec1fa73877d5e4c7c7"
                  egmiid : 444
                  fecha: "2019-11-04T18:15:32.673Z"  
                  sala_id: 7
                  codigo_evento: "AAA999"
                  __v: 0  
                # additionalProperties:
                #   type: integer
                #   format: int32
        '401':
          description: Invalid token
          content: {}
      security:
        - bearerAuth: []
    post:
      tags:
        - event
      summary: Envío de eventos
      operationId: addEvent
      requestBody:
        description: Agregado de un nuevo evento
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
          # application/xml:
          #   schema:
          #     $ref: '#/components/schemas/Event'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                example:
                  _id: "5dbc7cec1fa73877d5e4c7c7"
                  egmiid : 444
                  fecha: "2019-11-04T18:15:32.673Z"  
                  sala_id: 7
                  codigo_evento: "AAA999"
                  __v: 0  
        '405':
          description: Invalid input
          content: {}
        '401':
          description: Invalid token
          content: {}
      x-codegen-request-body-name: body
      security:
        - bearerAuth: []
  /user:
    post:
      tags:
        - user
      summary: Crear Usuario
      description: ''
      operationId: addUser
      requestBody:
        description: Crea un Usuario
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          # application/xml:
          #   schema:
          #     $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                example:
                  _id: "5dbc7cec1fa73877d5e4c7c7"
        '405':
          description: Invalid input
          content: {}
        '401':
          description: Invalid token
          content: {}
      x-codegen-request-body-name: body
      security:
        - bearerAuth: []
  /user/login:
    post:
      tags:
        - user
      summary: Login de usuario en el sistema
      operationId: loginUser
      requestBody:
        description: Logea a un Usuario
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
          # application/xml:
          #   schema:
          #     $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            # application/xml:
            #   schema:
            #     type: string
            #     items:
            #       $ref: '#/components/schemas/User'
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/User'
              example:
               token : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c3VhcmlvIjoiamRpZWd1ZXoiLCJpYXQiOjE1NzI4OTEyNDAsImV4cCI6MTU3Mjk3NzY0MH0.cQO0tmFMyWBZEiJ65pkmV8K9Rb5wFBLET03sk4w96xk"

        '400':
          description: Invalid username/password supplied
          content: {}
  /user/login/simetrico:
    post:
      tags:
        - user
      summary: Login de usuario con ecriptacion simetrica
      operationId: loginUserSymm
      requestBody:
        description: Logea a un usuario con ecriptacion simetrica
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginSymm'
          # application/xml:
          #   schema:
          #     $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            # application/xml:
            #   schema:
            #     type: string
            #     items:
            #       $ref: '#/components/schemas/User'
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/UserLoginSymm'
              example:
               token : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c3VhcmlvIjoiamRpZWd1ZXoiLCJpYXQiOjE1NzI4OTEyNDAsImV4cCI6MTU3Mjk3NzY0MH0.cQO0tmFMyWBZEiJ65pkmV8K9Rb5wFBLET03sk4w96xk"

        '400':
          description: Invalid username/password supplied
          content: {}
  /user/login/asimetrico:
    post:
      tags:
        - user
      summary: Login de usuario con ecriptacion asimetrica
      operationId: loginUserAsymm
      requestBody:
        description: Logea a un usuario con ecriptacion asimetrica
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginAsymm'
          # application/xml:
          #   schema:
          #     $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            # application/xml:
            #   schema:
            #     type: string
            #     items:
            #       $ref: '#/components/schemas/User'
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/UserLoginAsymm'
              example:
               token : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c3VhcmlvIjoiamRpZWd1ZXoiLCJpYXQiOjE1NzI4OTEyNDAsImV4cCI6MTU3Mjk3NzY0MH0.cQO0tmFMyWBZEiJ65pkmV8K9Rb5wFBLET03sk4w96xk"

        '400':
          description: Invalid username/password supplied
          content: {}        
components:
  schemas:
    User:
      type: object
      properties:
        nombre:
          type: string
        apellido:
          type: string
        usuario:
          type: string
          pattern: '^[a-zA-Z][A-Za-z0-9]+$'
        password:
          type: string
          pattern: '^(?=.*[A-Za-z])(?=.*\d)(?=.*[@$!%*#?&])[A-Za-z\d@$!%*#?&]{8,}$'
        status:
          type: boolean
        fecha_alta:
          type: string
          format: date-time
      xml:
        name: User
    Counter:
      type: object
      properties:
        egmid:
          type: integer
          format: int64
        ci:
          type: integer
          minimum: 0
          exclusiveMinimum: false
          maximum: 1000000000000000000
          exclusiveMaximum: true
        co:
          type: integer
          format: int64
          maximum: 1000000000000000000
          exclusiveMaximum: true
        jp:
          type: integer
          format: int64
          maximum: 1000000000000000000
          exclusiveMaximum: true
        jj:
          type: integer
          format: int64
          maximum: 1000000000000000000
          exclusiveMaximum: true
        drop:
          type: integer
          format: int64
          maximum: 1000000000000000000
          exclusiveMaximum: true
        cc:
          type: integer
          format: int64
          maximum: 1000000000000000000
          exclusiveMaximum: true
        fecha:
          type: string
          format: date-time
        sala_id:
          type: integer
          format: int64
      xml:
        name: User
    Event:
      type: object
      properties:
        egmid:
          type: integer
          format: int64
        fecha:
          type: string
          format: date-time
        sala_id:
          type: integer
          format: int64
        codigo_evento:
          type: string
          pattern: '^[a-fA-F0-9]+$'
        # xml:
        #   name: Event  
    UserLogin:
      type: object
      properties:
        usuario:
          type: string
          pattern: '^[a-zA-Z][A-Za-z0-9]+$'
        password:
          type: string
          pattern: '^(?=.*[A-Za-z])(?=.*\d)(?=.*[@$!%*#?&])[A-Za-z\d@$!%*#?&]{8,}$'
        # xml:
        #   name: UserLogin
    UserLoginSymm:
      type: object
      properties:
        usuario:
          type: string
          pattern: '^[a-zA-Z][A-Za-z0-9]+$'
        signature:
          type: string
    UserLoginAsymm:
      type: object
      properties:
        usuario:
          type: string
          pattern: '^[a-zA-Z][A-Za-z0-9]+$'
        signature:
          type: string                      
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    